% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analiza_wrazliwosci.R
\name{analiza_wrazliwosci_loo}
\alias{analiza_wrazliwosci_loo}
\title{Analiza Wrazliwosci Leave-One-Out}
\usage{
analiza_wrazliwosci_loo(
  macierz_decyzyjna,
  typy_kryteriow,
  metoda_wag = "equal",
  ...
)
}
\description{
Usuwa iteracyjnie jedna alternatywe i bada korelacje Spearmana.
}
