% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{calculate_dominance_ranking}
\alias{calculate_dominance_ranking}
\title{Teoria Dominacji dla Rankingu}
\usage{
calculate_dominance_ranking(rank_mat)
}
\description{
Funkcja pomocnicza do konsensusu.
}
\keyword{internal}
