% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_promethee.R
\name{.oblicz_preferencje_promethee}
\alias{.oblicz_preferencje_promethee}
\title{Internal Preference Calculator}
\usage{
.oblicz_preferencje_promethee(d, typ, q, p, s)
}
\description{
Oblicza wartości preferencji dla metody PROMETHEE.
}
\keyword{internal}
