% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.rozmyj_wektor}
\alias{.rozmyj_wektor}
\title{Wewnętrzna funkcja rozmywająca (Fuzzifier)}
\usage{
.rozmyj_wektor(wektor)
}
\description{
Zamienia liczbę rzeczywistą na Trójkątną Liczbę Rozmytą (TFN).
}
\keyword{internal}
