% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_promethee.R
\name{fuzzy_promethee}
\alias{fuzzy_promethee}
\title{Metoda Fuzzy PROMETHEE II}
\usage{
fuzzy_promethee(
  decision_matrix,
  criteria_type,
  preference_params,
  weights = NULL,
  bwm_best = NULL,
  bwm_worst = NULL
)
}
\arguments{
\item{decision_matrix}{Macierz (m × 3n) lub trójkątne liczby rozmyte (TFN).}

\item{criteria_type}{Wektor ("max" / "min") dla każdego kryterium.}

\item{preference_params}{Parametry funkcji preferencji.}

\item{weights}{Opcjonalny wektor numeryczny lub waga.}

\item{bwm_best}{Opcjonalny wektor BWM najlepsze_do_innych.}

\item{bwm_worst}{Opcjonalny wektor BWM inne_do_najlepszego.}
}
\value{
Object of class `fuzzy_promethee_res`
}
\description{
Implementuje metodę PROMETHEE II dla wielokryterialnej analizy decyzyjnej w środowisku rozmytym.
}
