% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_topsis.R
\name{fuzzy_topsis}
\alias{fuzzy_topsis}
\title{Metoda Fuzzy TOPSIS}
\usage{
fuzzy_topsis(
  decision_matrix,
  criteria_type,
  weights = NULL,
  bwm_best = NULL,
  bwm_worst = NULL,
  epsilon = 1e-09
)
}
\arguments{
\item{decision_matrix}{Macierz (m × 3n) lub TFN.}

\item{criteria_type}{Wektor ("max" / "min") dla każdego kryterium.}

\item{weights}{Opcjonalny wektor numeryczny lub waga.}

\item{bwm_best}{Opcjonalny wektor BWM najlepsze_do_innych.}

\item{bwm_worst}{Opcjonalny wektor BWM inne_do_najlepszego.}

\item{epsilon}{Mała stała numeryczna zapewniająca stabilność obliczeń
i zapobiegająca dzieleniu przez zero.}
}
\value{
Object of class `fuzzy_topsis_res`.
}
\description{
Implementuje metodę Fuzzy TOPSIS (Technique for Order Preference by Similarity
to Ideal Solution) dla wielokryterialnej analizy decyzyjnej z użyciem trójkątnych liczb rozmytych.
}
