% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_vikor.R
\name{fuzzy_vikor}
\alias{fuzzy_vikor}
\title{Metoda Fuzzy VIKOR}
\usage{
fuzzy_vikor(
  decision_matrix,
  criteria_type,
  strategy_weight = 0.5,
  weights = NULL,
  bwm_best = NULL,
  bwm_worst = NULL,
  epsilon = 1e-09
)
}
\arguments{
\item{decision_matrix}{Rozmyta macierz decyzyjna (m × 3n) zawierająca
trójkątne liczby rozmyte (TFN).}

\item{criteria_type}{Wektor typów kryteriów ("max" dla kryteriów zysków,
"min" dla kryteriów kosztów).}

\item{strategy_weight}{Wartość z przedziału [0,1] określająca znaczenie
użyteczności grupowej w metodzie VIKOR.}

\item{weights}{Opcjonalny wektor wag kryteriów w postaci rozmytej
(długość 3n).}

\item{bwm_best}{Opcjonalny wektor porównań najlepszego kryterium
względem pozostałych w metodzie BWM.}

\item{bwm_worst}{Opcjonalny wektor porównań pozostałych kryteriów
względem najgorszego w metodzie BWM.}

\item{epsilon}{Mała stała numeryczna zapewniająca stabilność obliczeń
i zapobiegająca dzieleniu przez zero.}
}
\value{
Object of class `fuzzy_vikor_res`.
}
\description{
Implementuje metodę Fuzzy VIKOR dla wielokryterialnej analizy decyzyjnej
w oparciu o ranking kompromisowy.
}
