% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{meta_ranking}
\alias{meta_ranking}
\title{Glowny Meta Ranking z wyposazeniem wagi dynamiczne}
\usage{
meta_ranking(
  macierz_decyzyjna,
  typy_kryteriow,
  metoda_wag = "equal",
  parametry_promethee = NULL,
  ...
)
}
\arguments{
\item{macierz_decyzyjna}{Rozmyta macierz decyzyjna.}

\item{typy_kryteriow}{Wektor ról kryteriów ("max" lub "min").}

\item{metoda_wag}{String wskazujący metodę wyznaczania wag ("equal", "sd", "gini", "bwm").}

\item{parametry_promethee}{Opcjonalna ramka danych (Type, q, p, s, Role) dla PROMETHEE.}

\item{...}{Parametry dodatkowe (np. 'najlepsze', 'najgorsze' dla BWM).}
}
\description{
Glowny Meta Ranking z wyposazeniem wagi dynamiczne
}
