% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{metoda_promethee}
\alias{metoda_promethee}
\title{Rozmyty PROMETHEE II (Prawidlowa Arytmetyka NEAT)}
\usage{
metoda_promethee(macierz, parametry_preferencji, wagi)
}
\arguments{
\item{macierz}{Rozmyta macierz decyzyjna.}

\item{parametry_preferencji}{Ramka danych z kolumnami: Type, q, p, s, Role.}

\item{wagi}{Wektor wag z funkcji oblicz_wagi.}
}
\description{
Rozmyty PROMETHEE II (Prawidlowa Arytmetyka NEAT)
}
