% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{oblicz_ranking_dominacji}
\alias{oblicz_ranking_dominacji}
\title{Konsensus Teoria Dominacji}
\usage{
oblicz_ranking_dominacji(macierz_rang)
}
\description{
Konsensus Teoria Dominacji
}
\keyword{internal}
