% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagi_helpers.R
\name{oblicz_wagi}
\alias{oblicz_wagi}
\title{Obliczanie wag obiektywnych i subiektywnych}
\usage{
oblicz_wagi(macierz_decyzyjna, metoda = "equal", ...)
}
\arguments{
\item{macierz_decyzyjna}{Rozmyta macierz (liczby trojkatne l, m, u).}

\item{metoda}{Wybrana metoda.}

\item{...}{Dodatkowe parametry dla BWM: nazwy_kryteriow, najlepsze, najgorsze.}
}
\description{
Wyznaczanie wag metodami: "equal", "sd", "gini", "bwm".
}
