% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fuzzy_topsis_res.R
\name{plot.fuzzy_topsis_res}
\alias{plot.fuzzy_topsis_res}
\title{Bubble plot for fuzzy TOPSIS results}
\usage{
\method{plot}{fuzzy_topsis_res}(x, ...)
}
\arguments{
\item{x}{Object of class `fuzzy_topsis_res`.}

\item{...}{Additional graphical parameters.}
}
\description{
Bubble plot for fuzzy TOPSIS results
}
