% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fuzzy_vikor_res.R
\name{plot.fuzzy_vikor_res}
\alias{plot.fuzzy_vikor_res}
\title{Bubble plot dla wyników metody Fuzzy VIKOR}
\usage{
\method{plot}{fuzzy_vikor_res}(x, ...)
}
\arguments{
\item{x}{Object of class `fuzzy_vikor_res`.}

\item{...}{Additional graphical parameters.}
}
\description{
Bubble plot dla wyników metody Fuzzy VIKOR
}
