% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot.rozmyty_topsis_wynik}
\alias{plot.rozmyty_topsis_wynik}
\title{Mapa Efektywności TOPSIS}
\usage{
\method{plot}{rozmyty_topsis_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy `rozmyty_topsis_wynik`.}

\item{...}{Dodatkowe argumenty.}
}
\description{
Oś X: Dystans od Anty-wzorca (D-). Oś Y: Dystans do Wzorca (D+).
}
