% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot.rozmyty_vikor_wynik}
\alias{plot.rozmyty_vikor_wynik}
\title{Mapa Strategiczna VIKOR}
\usage{
\method{plot}{rozmyty_vikor_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy `rozmyty_vikor_wynik`.}

\item{...}{Dodatkowe argumenty.}
}
\description{
Mapa Strategiczna VIKOR
}
