% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot_wagi}
\alias{plot_wagi}
\title{Wizualizacja Wag Kryteriów}
\usage{
plot_wagi(wagi_rozmyte, nazwy_kryteriow = NULL)
}
\arguments{
\item{wagi_rozmyte}{Wektor wag (format 3n: L, M, U dla każdego kryterium).}

\item{nazwy_kryteriow}{Opcjonalny wektor z nazwami kryteriów.}
}
\description{
Pokazuje defuzyfikowane wagi kryteriów wraz z rozpiętością rozmytą (L, U).
}
