% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{przygotuj_dane_mcda}
\alias{przygotuj_dane_mcda}
\title{Przygotowanie Danych do Rozmytej Analizy MCDA}
\usage{
przygotuj_dane_mcda(
  dane,
  skladnia,
  kolumna_alternatyw = NULL,
  funkcja_agregacji = mean
)
}
\arguments{
\item{dane}{Ramka danych (data frame) zawierająca surowe zmienne.}

\item{skladnia}{Ciąg znaków definiujący kryteria (np. "Koszt =~ k1 + k2").}

\item{kolumna_alternatyw}{Nazwa kolumny identyfikującej alternatywy.}

\item{funkcja_agregacji}{Funkcja używana do scalania opinii (domyślnie: mean).}
}
\value{
Macierz o wymiarach ($m \times 3n$).
}
\description{
Funkcja przekształca surowe dane w rozmytą macierz decyzyjną.
Skaluje zmienne do przedziału 1-9, agreguje opinie i dokonuje rozmycia.
}
