% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{rozmyty_multimoora}
\alias{rozmyty_multimoora}
\title{Rozmyta Metoda MULTIMOORA}
\usage{
rozmyty_multimoora(
  macierz_decyzyjna,
  typy_kryteriow,
  wagi = NULL,
  bwm_kryteria,
  bwm_najlepsze,
  bwm_najgorsze
)
}
\arguments{
\item{macierz_decyzyjna}{Macierz ($m \times 3n$).}

\item{typy_kryteriow}{Wektor znakowy ("max" dla zysku, "min" dla kosztu).}

\item{wagi}{(Opcjonalnie) Wektor wag.}

\item{bwm_kryteria}{(Opcjonalnie) Nazwy kryteriów dla BWM.}

\item{bwm_najlepsze}{(Opcjonalnie) Wektor Best-to-Others.}

\item{bwm_najgorsze}{(Opcjonalnie) Wektor Others-to-Worst.}
}
\value{
Obiekt klasy `rozmyty_multimoora_wynik`.
}
\description{
Implementacja metody Fuzzy MULTIMOORA. Składa się z:
1. Ratio System (RS)
2. Reference Point (RP)
3. Full Multiplicative Form (FMF)
Finalny ranking powstaje przez agregację Teorią Dominacji.
}
