% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_promethee.R
\name{rozmyty_promethee}
\alias{rozmyty_promethee}
\title{Rozmyta Metoda PROMETHEE II}
\usage{
rozmyty_promethee(
  macierz_decyzyjna,
  parametry_preferencji,
  wagi = NULL,
  bwm_kryteria,
  bwm_najlepsze,
  bwm_najgorsze
)
}
\arguments{
\item{macierz_decyzyjna}{Rozmyta macierz danych.}

\item{parametry_preferencji}{Ramka danych z kolumnami: Type, q, p, s, Role ("min"/"max").}

\item{wagi}{Wektor wag.}

\item{bwm_kryteria}{(Opcjonalnie) Nazwy kryteriów dla BWM.}

\item{bwm_najlepsze}{(Opcjonalnie) Wektor Best-to-Others.}

\item{bwm_najgorsze}{(Opcjonalnie) Wektor Others-to-Worst.}
}
\value{
Obiekt klasy `rozmyty_promethee_wynik`.
}
\description{
Implementacja metody Fuzzy PROMETHEE. Oblicza przepływy netto (Phi).
}
