% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mcda_input.R
\name{validate_mcda_input}
\alias{validate_mcda_input}
\title{Walidacja danych wejściowych}
\usage{
validate_mcda_input(decision_matrix, criteria_type)
}
\description{
Sprawdza poprawność danych wejściowych
}
\keyword{internal}
