% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{fuzzy_meta_ranking}
\alias{fuzzy_meta_ranking}
\title{Rozmyty Meta-Ranking (5 Metod)}
\usage{
fuzzy_meta_ranking(
  decision_mat,
  criteria_types,
  weights = NULL,
  preference_params = NULL,
  bwm_best = NULL,
  bwm_worst = NULL,
  lambda = 0.5,
  v = 0.5
)
}
\arguments{
\item{decision_mat}{Rozmyta macierz decyzyjna.}

\item{criteria_types}{Wektor typów kryteriów ("min", "max").}

\item{weights}{Wektor wag (opcjonalny).}

\item{preference_params}{Ramka parametrów dla PROMETHEE (opcjonalna).}

\item{bwm_best, bwm_worst}{Parametry BWM.}

\item{lambda}{Parametr WASPAS.}

\item{v}{Parametr VIKOR.}
}
\description{
Agreguje wyniki: VIKOR, TOPSIS, WASPAS, MULTIMOORA, PROMETHEE.
}
