% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_promethee.R
\name{fuzzy_promethee}
\alias{fuzzy_promethee}
\title{Fuzzy PROMETHEE II Method}
\usage{
fuzzy_promethee(
  decision_matrix,
  criteria_type,
  preference_params,
  weights = NULL,
  bwm_best = NULL,
  bwm_worst = NULL
)
}
\arguments{
\item{decision_matrix}{Matrix (m × 3n) of triangular fuzzy numbers.}

\item{criteria_type}{Character vector ("max" / "min") for each criterion.}

\item{preference_params}{Parameters of preference functions.}

\item{weights}{Optional numeric vector of weights.}

\item{bwm_best}{Optional BWM best-to-others vector.}

\item{bwm_worst}{Optional BWM others-to-worst vector.}
}
\value{
Object of class `fuzzy_promethee_res`
}
\description{
Implements PROMETHEE II for fuzzy multi-criteria decision making
}
