% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzzy_promethee_res.R
\name{plot.fuzzy_promethee_res}
\alias{plot.fuzzy_promethee_res}
\title{Bubble plot for fuzzy PROMETHEE results}
\usage{
\method{plot}{fuzzy_promethee_res}(x, ...)
}
\arguments{
\item{x}{Object of class `fuzzy_promethee_res`.}

\item{...}{Additional graphical parameters.}
}
\description{
Bubble plot for fuzzy PROMETHEE results
}
