% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mcda_input.R
\name{validate_mcda_input}
\alias{validate_mcda_input}
\title{MCDA Input Validation}
\usage{
validate_mcda_input(decision_matrix, criteria_type)
}
\description{
Checks correctness of MCDA inputs
}
\keyword{internal}
